Added localisations, improved upon the cutting mechanism, added an fps counter and a logger to measure the average fps over the lifespan of the application for statistical purposes.

+++ Pyro.Nc/Simulation/CutResult.cs
@@ -2,11 +2,11 @@ namespace Pyro.Nc.Simulation
 {
     public struct CutResult
     {
-        public long TotalTime;
+        public double TotalTime;
         public long TotalVerticesCut;
         public bool Threw;
 
-        public CutResult(long totalTime, long totalVerticesCut)
+        public CutResult(double totalTime, long totalVerticesCut)
         {
             TotalTime = totalTime;
             TotalVerticesCut = totalVerticesCut;
